Add SQL upgrade statement for user table

+++ config/db/mysql-create-tables.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS user (
   secret varchar(128),
   loginattempts integer,
   tmpblocktimestamp BIGINT,
-  tmpblockattempts BIGINT,
+  tmpblockattempts integer,
   blocked tinyint(1),
   notificationtype varchar(10),
   notificationaddress varchar(64)
